changes by Mufaddal

+++ src/app/Shared-Module/Services/request-handler.service.ts
@@ -116,10 +116,14 @@ export class RequestHandlerService {
 
   // Handle token expiration
   private handleExpiredToken(error: HttpErrorResponse): Observable<HttpEvent<HttpErrorResponse>> {
-    this.authService.clearUserInfo();
-    localStorage.clear();
-    sessionStorage.clear();
-    this.router.navigate(['/user/login']);
+    let errors = (typeof error.error === 'object') ? error.error : {message:''};
+    if (error.status === 401 || error.status === 403 && ((error.error && typeof error.error  === 'object') && errors.message.includes('token'))) {
+      this.authService.clearUserInfo();
+      localStorage.clear();
+      sessionStorage.clear();
+      this.router.navigate(['/user/login']);
+      return throwError(() => 'Session expired') as Observable<HttpEvent<HttpErrorResponse>>;
+    }
     return throwError(() => error) as Observable<HttpEvent<HttpErrorResponse>>;
   }
